requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
whitenoise==6.4.0
wsproto==1.2.0
